# JIT: Remove BB limit from importer_vectorization