@EgorBo 2022-03-12
> > surprisingly, it makes code even smaller.
> 
> Thanks, Egor. I was surprised in the repro when Eric showed it to me last night that even AsSpan wasn't being inlined. Separate from this change, have we done an experiment around what things would look like if we always, unconditionally inlined AsSpan and span's ctors? I wonder if there's a set of critical, core, tiny APIs where we end up falling off a cliff due to inlining budget and we'd be better off with an internal MethodImplOptions.ForceInliningNoReallyThisNeedsToBeInlined setting.

We plan to address it eventually with partial pre-scan to be able to go over-budget, I think there must be already an issue filed for it.