@kasperk81 2022-03-13
> 652 total methods with Code Size differences (428 improved, 224 regressed), 1 unchanged.

what caused the regressions?