@EgorBo 2022-03-13
> > 652 total methods with Code Size differences (428 improved, 224 regressed), 1 unchanged.
> 
> what caused the regressions?

It seems like it's always a size improvement to unroll string.Equals for spans, but it's often a size regression for string objects.
However, I checked few size regressions and all of them were "performance" improvements so I leave it as is.