@jkotas 2022-03-13
> I suspect it's DST difference

Opened https://github.com/dotnet/runtime/issues/66555